fix(types): correct NDArray dtype from float64 to int32
- _remove_short_events now correctly uses int32 throughout
- Fixes type mismatch in binary array processing
- Ensures consistency with astype(np.int32) calls

+++ src/brain_go_brrr/infra/ml_models/seizure_transformer_utils.py
@@ -151,7 +151,7 @@ class SeizurePostProcessor:
 
         return binary.astype(np.int32)
 
-    def _remove_short_events(self, binary: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
+    def _remove_short_events(self, binary: npt.NDArray[np.int32]) -> npt.NDArray[np.int32]:
         """Remove seizure events shorter than min_duration."""
         from scipy.ndimage import label
